requirements.txt
torch
scikit-learn
pandas
gradio
PyPDF2
requests
python-dotenv
